Adding - relationship in schema - methods for better testing - better schema - null checking

+++ src/main/java/com/example/transactionchallenge/controller/dto/TransactionResponse.java
@@ -3,6 +3,6 @@ package com.example.transactionchallenge.controller.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record TransactionResponse(
-        @JsonProperty("account_id") int accountId,
+        @JsonProperty("account_id") Long accountId,
         @JsonProperty("operation_type_id") int operationTypeId,
         Double amount) {}
